Moved translation auxiliaries like pencil construction and cell flagging to a separate file. Passes transtest2 in cpu mode. GPU mode broken atm.
